requirements.txt
python-tornado
python-daemon
python-beautifulsoup
pianobar
aplay
espeak
sox
libttspico-utils
